Store libev callbacks in an internal buffer and return them. Tests pass!
NIO::Selectors keep an internal buffer of received events completely in
C which it uses while the GIL is free. This set of changes populates
that buffer, waits for the GIL to be reacquired, then creates a new Ruby
array object out of the buffer's contents.

It also implements the timeout for NIO::Selector#select

This is the long pole and it's cleared! Now down to one failing test. All
that remains now is the wakeup mechanism