NIO::Selector#select_each support for pure Ruby and JRuby
The present NIO::Selector#select API, following in the spirit of Kernel.select, allocates a new array every time you call it. This is wasteful and unnecessary. An iterator API can avoid allocating memory and reuse a buffer between calls.

Unfortunately, C is trolling me, so this commit only contains the JRuby and pure Ruby versions.