Fix several issues with Java NIO semantics
Java NIO is... kind of quirky compared to what I'm used to. It provides
separate interest types for sockets that are accepting or connecting, in
contrast to POSIX which uses reading and writing for those purposes.

Apparently Java NIO wants you to clear the selectedKeys yourself. I
guess this makes sense in a garbage collected language, however I'm used
to this sort of buffer being fixed and automatically cleared between
calls.