Refactor NIO::Selector#select and #select_each
There's a lot of repetition between these two methods. This refactors
some of it into a separate method. There's still a lot going on though.

This lead to an uncaught bug: the selectedKeys weren't being cleared in
NIO::Selector#select, only #select_each, which lead to a bug.

I'm not too worried as I'll probably be translating this code to Java
soon :/