Spec for reregistering IO objects with the same selector
Scumbag Java NIO strikes again! Apparently NIO tries to batch all
registration/deregistration operations, probably to take advantage of
system calls that do the same. Yet they expose this as an end user
concern. WTF Java people, seriously?

nio4r will, again, NOT be following in the Java pattern. The NIO
behavior is considered a BUG, and the solution will be to fix it. But
for now, the current implementation is broken! Ohnoes!