Allow reregistration of deregistered IO objects
NIO doesn't let you do this, ostensibly to improve performance to the
low-level system calls. Should this be a concern of the user of the NIO
API? I guess if you're roflscale.

For the typical Ruby user, the answer is NO. This code batches key
cancellation automatically inside a per-selector hash map, and reuses
SelectionKeys if they're requested again before the next selection.