Separate Selector#select return value for timeout/wakeup
wakeup: false
timeout: nil

For pure implementation we return nil if Kernel.select returns nil,
this does not handle spurious wakeups properly (if they can happen at all).

For C implementation we return nil if the timed_out flag was set as part
of the timeout timer, or if the busy loop determines we have timed out.

For Java implementation we calculate time spent waiting for select(),
if it is longer than the specified timeout we assume that we have timed
out.