Don't build the extension on Windows
The extension can be build on Windows but it doesn't work.
All connection related tests fail.
So it's a waste of time to build the extension, which isn't used.

This also removes the patching of the makefile, because it isn't
necessary on any recent ruby version, but breaks the extension,
so that it can't be loaded by ruby.
So this should be removed anyway.

I tested this on ruby-2.2-x86 and 2.4-x64.

See also #161