Style consistency using clang-format
The styles in the different files weren't as consistent as one might prefer. This fixes this issue.

Also, this makes it easier for collaborators to author PRs, by making use of an automated styling approach.

The clang-format settings took what I could understand as the prevailing style (4 spaces indentation, when to brace, etc') and I applied the nearest match. Aside for macro and include statements that I don't know how to make clang-format indent, this is a good-enough collaborative setting.