allow missing `devkit`
As per the comments in the build file, windows does not actually use
any native extension code.

Not sure why the require was added for `devkit`.  In the interest of
compatibility simply guard the require with a rescue for the `LoadError`.

If the Ruby environment was custom built the `devkit` class will not
always be available. In some cases such as more recent builds for
Ruby 3.0.x the devkit can utilize an MSYS2 environment without needing
the `devkit` class.